refactor: ナビゲーションタブバーのコードを移動した
