+++ lib/component/navigation_bar.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+import 'package:routine_app/view/home/home_view.dart';
+import 'package:routine_app/view/setting/setting_view.dart';
+
+class Navigation extends StatefulWidget {
+  const Navigation({super.key});
+
+  @override
+  State<Navigation> createState() => _NavigationState();
+}
+
+class _NavigationState extends State<Navigation> {
+  int currentPageIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      bottomNavigationBar: NavigationBar(
+        onDestinationSelected: (int index) {
+          setState(() {
+            currentPageIndex = index;
+          });
+        },
+        indicatorColor: Colors.amber,
+        selectedIndex: currentPageIndex,
+        destinations: const <Widget>[
+          NavigationDestination(
+            selectedIcon: Icon(Icons.home),
+            icon: Icon(Icons.home_outlined),
+            label: 'ホーム',
+          ),
+          NavigationDestination(
+            icon: Icon(Icons.settings),
+            label: '設定',
+          ),
+        ],
+      ),
+      body: <Widget>[
+        const HomeView(),
+        const SettingView(),
+      ][currentPageIndex],
+    );
+  }
+}
