+++ lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:routine_app/view/home/home_view.dart';
-import 'package:routine_app/view/setting/setting_view.dart';
+import 'package:routine_app/component/navigation_bar.dart';
 
 void main() {
   runApp(const MyApp());
@@ -13,47 +12,6 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return const MaterialApp(
       home: Navigation(),
-    );
-  }
-}
-
-class Navigation extends StatefulWidget {
-  const Navigation({super.key});
-
-  @override
-  State<Navigation> createState() => _NavigationState();
-}
-
-class _NavigationState extends State<Navigation> {
-  int currentPageIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      bottomNavigationBar: NavigationBar(
-        onDestinationSelected: (int index) {
-          setState(() {
-            currentPageIndex = index;
-          });
-        },
-        indicatorColor: Colors.amber,
-        selectedIndex: currentPageIndex,
-        destinations: const <Widget>[
-          NavigationDestination(
-            selectedIcon: Icon(Icons.home),
-            icon: Icon(Icons.home_outlined),
-            label: 'ホーム',
-          ),
-          NavigationDestination(
-            icon: Icon(Icons.settings),
-            label: '設定',
-          ),
-        ],
-      ),
-      body: <Widget>[
-        const HomeView(),
-        const SettingView(),
-      ][currentPageIndex],
     );
   }
 }
